+++ CHANGES.rst
@@ -6,7 +6,7 @@ There's a frood who really knows where his towel is.
 2.13b2 (unreleased)
 ^^^^^^^^^^^^^^^^^^^
 
-- Add a way to extend default metadata using adapters
+- Add a way to extend and/or override Open Graph metadata using adapters.
   [erral]
 
 - Complete Basque translation.